Refactor backend code structure
