+++ routers/order.py
@@ -1,0 +1,4 @@
+from fastapi import APIRouter
+
+router = APIRouter()
+
